Separated everything
Also added some stuff.
